+++ src/main/java/MusicApp/repositories/SongRepository.java
@@ -3,5 +3,5 @@ package MusicApp.repositories;
 import MusicApp.models.Song;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SongRepository extends CrudRepository<Song, Integer > {
+public interface SongRepository extends CrudRepository<Song, Long> {
 }
